requirements.txt
-i https://pypi.org/simple
dj-database-url==0.5.0
django-redis==4.10.0
django==2.1.5
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
pygments==2.3.1
python-dotenv==0.10.1
pytz==2018.9
redis==3.0.1
